backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.7.0
reportlab>=4.2.0
python-dotenv>=1.0.0
pydantic>=2.7.0
typing-extensions>=4.8.0